+++ docs/changelog.md
@@ -2,6 +2,15 @@
 
 <!--next-version-placeholder-->
 
+## v0.2.0 (09/2023)
+
+- Added [authentication](api/authentication.md) to allow authenticated requests.
+- Default to version 2 of the REST API.
+- Added new methods for uploading, updating, and deleting assets.
+- Added new methods obtaining information on taxonomies.
+- Updated the documentation significantly.
+
+
 ## v0.1.0 (06/12/2023)
 
 - First release of `aiondemand`!